Handle batches that failed to send
A failed batch puts the aggregator in a stopped state. It no longer
sends batches and no longer accepts new submits.
The admin can call resume() to transition to the resuming state.
The aggregator then starts to send batches made up of the tx requests
that were in failed batches. When those are all send then it
transitions to the normal running state.